pyproject.toml
[project]
name = "langchain-course-rag-gist"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.12.0",
    "isort>=7.0.0",
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.6",
    "langchain-pinecone>=0.2.13",
    "langchainhub>=0.1.21",
    "python-dotenv>=1.2.1",
]
